[TOPPRA] Split into header and definition file.
