+++ plugins/toppra.cc
@@ -15,6 +15,8 @@
 // received a copy of the GNU Lesser General Public License along with
 // hpp-core. If not, see <http://www.gnu.org/licenses/>.
 
+#include "toppra.hh"
+
 #include <pinocchio/multibody/model.hpp>
 
 #include <hpp/pinocchio/device.hh>
@@ -80,127 +82,109 @@ private:
 
 namespace pathOptimization
 {
-class TOPPRA;
-typedef boost::shared_ptr<TOPPRA> TOPPRAPtr_t;
-
-class TOPPRA : public PathOptimizer
+PathVectorPtr_t TOPPRA::optimize(const PathVectorPtr_t &path)
 {
-  public:
-    static TOPPRAPtr_t create(const Problem &p)
+  const size_type solver = problem().getParameter(PARAM_HEAD "solver").intValue();
+  const value_type effortScale = problem().getParameter(PARAM_HEAD "effortScale").floatValue();
+  const value_type velScale = problem().getParameter(PARAM_HEAD "velocityScale").floatValue();
+
+  using pinocchio::Model;
+  const Model &model = problem().robot()->model();
+
+  using namespace toppra::constraint;
+
+  // Create the TOPPRA constraints
+  auto torqueConstraint = std::make_shared<jointTorque::Pinocchio<Model> > (model); // No friction
+  torqueConstraint->lowerBounds(effortScale * torqueConstraint->lowerBounds());
+  torqueConstraint->upperBounds(effortScale * torqueConstraint->upperBounds());
+
+  toppra::LinearConstraintPtrs v{
+    std::make_shared<LinearJointVelocity>(
+        -velScale * model.velocityLimit, velScale * model.velocityLimit)
+      //, std::make_shared<LinearJointAcceleration>(
+      //- 10 * model.velocityLimit, 10 * model.velocityLimit)
+      , torqueConstraint
+  };
+
+  PathVectorPtr_t flatten_path = PathVector::create(path->outputSize(), path->outputDerivativeSize());
+  path->flatten(flatten_path);
+  PathVectorPtr_t res = PathVector::create(path->outputSize(), path->outputDerivativeSize());
+  for(auto idx_subpath = 0ul; idx_subpath < flatten_path->numberPaths(); ++idx_subpath)
+  {
+    const auto subpath = flatten_path->pathAtRank(idx_subpath);
+    std::shared_ptr<PathWrapper> pathWrapper (std::make_shared<PathWrapper>(subpath));
+
+    toppra::algorithm::TOPPRA algo(v, pathWrapper);
+    algo.setN((int)problem().getParameter(PARAM_HEAD "N").intValue());
+    switch(solver) {
+      default:
+        hppDout (error, "Solver " << solver << " does not exists. Using GLPK");
+      case 0:
+        algo.solver(std::make_shared<toppra::solver::GLPKWrapper>());
+        break;
+      case 1:
+        algo.solver(std::make_shared<toppra::solver::qpOASESWrapper>());
+        break;
+    }
+    auto ret_code = algo.computePathParametrization();
+    if (ret_code != toppra::ReturnCode::OK)
     {
-      return TOPPRAPtr_t(new TOPPRA(p));
+      std::stringstream ss;
+      ss << "TOPPRA failed, returned code: " << static_cast<int>(ret_code) << std::endl;
+      throw std::runtime_error(ss.str());
+    }
+    const auto out_data = algo.getParameterizationData();
+
+    // forward integration of time parameterization (trapezoidal integration)
+    assert(out_data.gridpoints.size() == out_data.parametrization.size());
+    const size_t num_pts = out_data.gridpoints.size();
+    auto sd = toppra::Vector(num_pts);
+    for (auto i = 0ul; i < num_pts; ++i)
+    {
+      sd[i] = std::sqrt(std::max(out_data.parametrization[i], 0.));
     }
 
-  PathVectorPtr_t optimize(const PathVectorPtr_t &path)
-  {
-    const size_type solver = problem().getParameter(PARAM_HEAD "solver").intValue();
-    const value_type effortScale = problem().getParameter(PARAM_HEAD "effortScale").floatValue();
-    const value_type velScale = problem().getParameter(PARAM_HEAD "velocityScale").floatValue();
-
-    using pinocchio::Model;
-    const Model &model = problem().robot()->model();
-
-    using namespace toppra::constraint;
-
-    // Create the TOPPRA constraints
-    auto torqueConstraint = std::make_shared<jointTorque::Pinocchio<Model> >
-      (model, toppra::Vector::Constant(model.nv, 0.)); // No friction
-    torqueConstraint->lowerBounds(effortScale * torqueConstraint->lowerBounds());
-    torqueConstraint->upperBounds(effortScale * torqueConstraint->upperBounds());
-
-    toppra::LinearConstraintPtrs v{
-        std::make_shared<LinearJointVelocity>(
-            -velScale * model.velocityLimit, velScale * model.velocityLimit)
-        //, std::make_shared<LinearJointAcceleration>(
-            //- 10 * model.velocityLimit, 10 * model.velocityLimit)
-        , torqueConstraint
-            };
-  
-    PathVectorPtr_t flatten_path = PathVector::create(path->outputSize(), path->outputDerivativeSize());
-    path->flatten(flatten_path);
-    PathVectorPtr_t res = PathVector::create(path->outputSize(), path->outputDerivativeSize());
-    for(auto idx_subpath = 0ul; idx_subpath < flatten_path->numberPaths(); ++idx_subpath)
+    auto t = toppra::Vector(num_pts);
+    t[0] = 0.; // start time is 0
+    for (auto i = 1ul; i < num_pts; ++i)
     {
-      const auto subpath = flatten_path->pathAtRank(idx_subpath);
-      std::shared_ptr<PathWrapper> pathWrapper (std::make_shared<PathWrapper>(subpath));
-
-      toppra::algorithm::TOPPRA algo(v, pathWrapper);
-      algo.setN((int)problem().getParameter(PARAM_HEAD "N").intValue());
-      switch(solver) {
-        default:
-          hppDout (error, "Solver " << solver << " does not exists. Using GLPK");
-        case 0:
-          algo.solver(std::make_shared<toppra::solver::GLPKWrapper>());
-          break;
-        case 1:
-          algo.solver(std::make_shared<toppra::solver::qpOASESWrapper>());
-          break;
-      }
-      auto ret_code = algo.computePathParametrization();
-      if (ret_code != toppra::ReturnCode::OK)
-      {
-        std::stringstream ss;
-        ss << "TOPPRA failed, returned code: " << static_cast<int>(ret_code) << std::endl;
-        throw std::runtime_error(ss.str());
-      }
-      const auto out_data = algo.getParameterizationData();
-
-      // forward integration of time parameterization (trapezoidal integration)
-      assert(out_data.gridpoints.size() == out_data.parametrization.size());
-      const size_t num_pts = out_data.gridpoints.size();
-      auto sd = toppra::Vector(num_pts);
-      for (auto i = 0ul; i < num_pts; ++i)
-      {
-        sd[i] = std::sqrt(std::max(out_data.parametrization[i], 0.));
-      }
-
-      auto t = toppra::Vector(num_pts);
-      t[0] = 0.; // start time is 0
-      for (auto i = 1ul; i < num_pts; ++i)
-      {
-        const auto sd_avg = (sd[i - 1] + sd[i]) * 0.5;
-        const auto ds = out_data.gridpoints[i] - out_data.gridpoints[i - 1];
-        assert(sd_avg > 0.);
-        const auto dt = ds / sd_avg;    
-        t[i] = t[i - 1] + dt;
-      }
-
-      // time parameterization based on hermite cubic spline interpolation
-      constexpr int order = 3;
-      typedef timeParameterization::PiecewisePolynomial<order> timeparam;
-      auto params = timeparam::ParameterMatrix_t(order + 1, num_pts - 1);
-      const auto& s = out_data.gridpoints;
-      for (auto i = 1ul; i < num_pts; ++i)
-      {
-        const auto inv_dt = 1./(t[i] - t[i-1]);
-        const auto inv_dt2 = inv_dt*inv_dt;
-        const auto inv_dt3 = inv_dt2*inv_dt;
-        const auto t_p = t[i-1];
-        const auto t_p2 = t_p*t_p;
-        const auto t_p3 = t_p2*t_p;
-        const auto ds = (s[i] - s[i-1]);
-        const auto b = (2*sd[i-1] + sd[i])*inv_dt;
-        const auto c = (sd[i-1] + sd[i])*inv_dt2;
-        params(0, i-1) = 2*t_p3*ds*inv_dt3 + 3*t_p2*ds*inv_dt2 - t_p3*c -t_p2*b - t_p*sd[i-1] + s[i-1];
-        params(1, i-1) = -6*t_p2*ds*inv_dt3 - 6*t_p*ds*inv_dt2 + 3*t_p2*c + 2*t_p*b + sd[i-1];
-        params(2, i-1) = 6*t_p*ds*inv_dt3 + 3*ds*inv_dt2 - 3*t_p*c - b;
-        params(3, i-1) = -2*ds*inv_dt3 + c;
-      }
-
-      subpath->timeParameterization(TimeParameterizationPtr_t(new timeparam(params, t)),
-                              interval_t(t[0], t[num_pts - 1]));
-      res->appendPath(subpath);
-
+      const auto sd_avg = (sd[i - 1] + sd[i]) * 0.5;
+      const auto ds = out_data.gridpoints[i] - out_data.gridpoints[i - 1];
+      assert(sd_avg > 0.);
+      const auto dt = ds / sd_avg;    
+      t[i] = t[i - 1] + dt;
     }
 
-    return res;
+    // time parameterization based on hermite cubic spline interpolation
+    constexpr int order = 3;
+    typedef timeParameterization::PiecewisePolynomial<order> timeparam;
+    auto params = timeparam::ParameterMatrix_t(order + 1, num_pts - 1);
+    const auto& s = out_data.gridpoints;
+    for (auto i = 1ul; i < num_pts; ++i)
+    {
+      const auto inv_dt = 1./(t[i] - t[i-1]);
+      const auto inv_dt2 = inv_dt*inv_dt;
+      const auto inv_dt3 = inv_dt2*inv_dt;
+      const auto t_p = t[i-1];
+      const auto t_p2 = t_p*t_p;
+      const auto t_p3 = t_p2*t_p;
+      const auto ds = (s[i] - s[i-1]);
+      const auto b = (2*sd[i-1] + sd[i])*inv_dt;
+      const auto c = (sd[i-1] + sd[i])*inv_dt2;
+      params(0, i-1) = 2*t_p3*ds*inv_dt3 + 3*t_p2*ds*inv_dt2 - t_p3*c -t_p2*b - t_p*sd[i-1] + s[i-1];
+      params(1, i-1) = -6*t_p2*ds*inv_dt3 - 6*t_p*ds*inv_dt2 + 3*t_p2*c + 2*t_p*b + sd[i-1];
+      params(2, i-1) = 6*t_p*ds*inv_dt3 + 3*ds*inv_dt2 - 3*t_p*c - b;
+      params(3, i-1) = -2*ds*inv_dt3 + c;
+    }
+
+    subpath->timeParameterization(TimeParameterizationPtr_t(new timeparam(params, t)),
+        interval_t(t[0], t[num_pts - 1]));
+    res->appendPath(subpath);
+
   }
 
-protected:
-  TOPPRA(const Problem &problem) : PathOptimizer(problem)
-  {
-  }
-};
+  return res;
+}
 
 HPP_START_PARAMETER_DECLARATION(TOPPRA)
 Problem::declareParameter(ParameterDescription (Parameter::FLOAT,
